Fixed buggy reset
Intro sequence

An introductory sequence where the players circumvent
the TRON board and
then begin the actual game. The players are placed in
opposite corners
at the begining(changes to the entityManager) and are
given a special
introUpdate function that overrides and user input or
AI, and is done
for the length of two board edges, manged by the counter
introCount.

Resetting and Colliding

Player resets upon colliding with another
player/wall. The board is
cleared. The player´s velocities is the same
as their initial velocities
and inheritance from the player's ancestor
is prevented by clearing the
appropriate states in keys. (This does not
work properly!). The
spatialManager is changed in order for empty
vertices not to be operated
on when resetting.